Don't delete config files in build dir

+++ build.sh
@@ -4,12 +4,14 @@ build() {
         if [[ ! -e build ]]; then
                 mkdir build
         else
-                rm -rf build/*
+                rm -rf build/irc-client
         fi
 
         cd cmd
         go build -o ${PROJECT_DIR}/build/irc-client
-        cp ${PROJECT_DIR}/misc/UserConfigs.json ${PROJECT_DIR}/build
+        if [[ ! -e ${PROJECT_DIR}/misc/UserConfigs.json ]]; then
+                cp ${PROJECT_DIR}/misc/UserConfigs.json ${PROJECT_DIR}/build
+        fi
 }
 
 build
